Major refactor of radar-auth
No longer implemented as a JAX-RS filter. Current implementation allows client of this library to easily implement their own filter that adds the decoded JWT to the context.
