+++ radar-auth/build.gradle
@@ -35,23 +35,15 @@ publishing {
 }
 
 dependencies {
-    api group: 'com.squareup.okhttp3', name: 'okhttp', version: okhttpVersion
-
     compile group: 'ch.qos.logback', name: 'logback-classic', version: '1.2.2'
 
     compile group: 'com.auth0', name: 'java-jwt', version: '3.2.0'
-
-    compile group: 'javax.ws.rs', name: 'javax.ws.rs-api', version: '2.1-m05'
-    compile group: 'javax.annotation', name: 'javax.annotation-api', version: '1.2-b02'
-    compile group: 'org.radarcns', name: 'managementportal-client', version: '0.0.1-SNAPSHOT'
 
     implementation group: 'com.fasterxml.jackson.core' , name: 'jackson-databind' , version: jacksonVersion
     implementation group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-yaml', version: jacksonVersion
 
     testCompile group: 'junit', name: 'junit', version: junitVersion
     testCompile group: 'com.github.tomakehurst', name: 'wiremock', version: '2.5.1'
-    testCompile group: 'org.mockito', name: 'mockito-core', version: '2.8.47'
-    testCompile 'org.glassfish.jersey.core:jersey-common:2.22.2'
     testCompile 'com.github.stefanbirkner:system-rules:1.16.0'
 }
