+++ src/main/resources/config/application-dev.yml
@@ -87,7 +87,7 @@ liquibase:
 #        keyAlias: ManagementPortal
 # ===================================================================
 server:
-    port: 8080
+    port: 8089
 
 # ===================================================================
 # JHipster specific properties
